refactor: move endpoints to Routes directory
